Changed point in time when keystore password is read from system property
Previously, the value of the system property javax.net.ssl.keyStorePassword
was read once at the instantiation time of the SSLManager class. In contrast,
the value of the system property javax.net.ssl.keyStore is read for
each invocation of SSLManager#getKeyStore.

Now, both system properties are read when SSLManager#getKeyStore is invoked.